Add files via upload
In this analysis, the focus begins with thorough data preparation and exploratory data analysis (EDA). This phase is crucial for identifying patterns, anomalies, and insights within the dataset. Afterward, both unsupervised and supervised machine learning techniques are applied to uncover patterns, categorize tumors, and predict numerical features. This combined approach leverages the full potential of the data, leading to actionable insights. Emphasis is placed on ensuring transparency and reproducibility in the methodology. Additionally, ethical considerations, especially in healthcare applications, are carefully addressed to ensure responsible use of machine learning.